[fabric] Update add_peer and add_orderer to 2.5.4 Primary Changes -------------- 1.add_peer updated for version 2.5.4 2.add_orderer updated for version 2.5.4
Modifications
-----------------------
platforms/hyperledger-fabric/charts/fabric-orderernode
platforms/hyperledger-fabric/charts/fabric-osnadmin-channel-create
platforms/hyperledger-fabric/configuration/add-orderer.yaml
platforms/hyperledger-fabric/configuration/add-peer.yaml
platforms/hyperledger-fabric/configuration/roles/create/new_orderer/create_appchannel_block/templates/update_channel_script.tpl
platforms/hyperledger-fabric/configuration/roles/create/orderers/tasks/main.yaml
platforms/hyperledger-fabric/configuration/roles/create/osnchannels/tasks/valuefile.yaml
platforms/hyperledger-fabric/configuration/roles/create/peers/tasks/main.yaml
platforms/hyperledger-fabric/configuration/roles/create/peers/tasks/nested_main.yaml
platforms/hyperledger-fabric/configuration/roles/helm_component/templates/orderernode.tpl
platforms/hyperledger-fabric/configuration/roles/helm_component/templates/osn_create_channel_job.tpl

fixes #2430

+++ platforms/hyperledger-fabric/configuration/roles/create/new_orderer/create_appchannel_block/templates/update_channel_script.tpl
@@ -3,22 +3,34 @@
 set -x
 
 CURRENT_DIR=${PWD}
+NETWORK_VERSION="{{ version }}"
 
 echo "installing jq "
 apt-get install -y jq
-echo "installing configtxlator"
-mkdir temp
-cd temp/
-wget https://github.com/hyperledger/fabric/releases/download/v{{ version }}/hyperledger-fabric-{{ os }}-{{ arch }}-{{ version }}.tar.gz
-tar -xvf hyperledger-fabric-{{ os }}-{{ arch }}-{{ version }}.tar.gz
-mv bin/configtxlator ../
-cd ../
-rm -r temp
+if [ "$NETWORK_VERSION" != "2.5.4" ]; then
+    echo "installing configtxlator"
+    mkdir temp
+    cd temp/
+    wget https://github.com/hyperledger/fabric/releases/download/v{{ version }}/hyperledger-fabric-{{ os }}-{{ arch }}-{{ version }}.tar.gz
+    tar -xvf hyperledger-fabric-{{ os }}-{{ arch }}-{{ version }}.tar.gz
+    mv bin/configtxlator ../
+    cd ../
+    rm -r temp
+fi
+
 echo "converting the channel_config_block.pb to channel_config.json using configtxlator and jq"
 configtxlator proto_decode --input {{ channel_name }}_config_block.pb --type common.Block | jq .data.data[0].payload.data.config > {{ channel_name }}_config.json
 echo "adding new organization crypto material from config.json to the channel_config.json to make channel_modified_config.json"
-jq --argjson a "$(cat ./orderer)" '.channel_group.values.OrdererAddresses.value.addresses += $a' {{ channel_name }}_config.json > {{ channel_name }}_modified_intermediate_config.json
-jq --argjson a "$(cat ./orderer-tls)" '.channel_group.groups.Orderer.values.ConsensusType.value.metadata.consenters += $a' {{ channel_name }}_modified_intermediate_config.json > {{ channel_name }}_modified_config.json
+if [ "$NETWORK_VERSION" != "2.5.4" ]; then
+    echo "version 2.2.2++++"
+    jq --argjson a "$(cat ./orderer)" '.channel_group.values.OrdererAddresses.value.addresses += $a' {{ channel_name }}_config.json > {{ channel_name }}_modified_intermediate_config.json
+    jq --argjson a "$(cat ./orderer-tls)" '.channel_group.groups.Orderer.values.ConsensusType.value.metadata.consenters += $a' {{ channel_name }}_modified_intermediate_config.json > {{ channel_name }}_modified_config.json
+else
+    echo "version 2.5.4++++"
+    jq --argjson a "$(cat ./orderer)" '.channel_group.values.OrdererAddresses.value.addresses += $a' {{ channel_name }}_config.json > {{ channel_name }}_modified_intermediate_address_config.json
+    jq --argjson a "$(cat ./orderer)" '.channel_group.groups.Orderer.groups.{{ component_name }}MSP.values.Endpoints.value.addresses += $a' {{ channel_name }}_modified_intermediate_address_config.json > {{ channel_name }}_modified_intermediate_endpoints_config.json
+    jq --argjson a "$(cat ./orderer-tls)" '.channel_group.groups.Orderer.values.ConsensusType.value.metadata.consenters += $a' {{ channel_name }}_modified_intermediate_endpoints_config.json > {{ channel_name }}_modified_config.json
+fi
 echo "converting the channel_config.json and channel_modified_config.json to .pb files"
 configtxlator proto_encode --input {{ channel_name }}_config.json --type common.Config --output {{ channel_name }}_config.pb
 configtxlator proto_encode --input {{ channel_name }}_modified_config.json --type common.Config --output {{ channel_name }}_modified_config.pb
